modified:   __pycache__/create_par.cpython-39.pyc 	modified:   __pycache__/edit_page.cpython-39.pyc 	new file:   __pycache__/edit_paragraph.cpython-39.pyc 	modified:   __pycache__/move_tags.cpython-39.pyc 	modified:   create_par.py 	modified:   edit_page.py 	new file:   edit_paragraph.py 	modified:   html_files/index.html 	modified:   saves/selected_id.txt

+++ edit_page.py
@@ -11,12 +11,27 @@ def edit():
          python = sys.executable
          os.execl(python, python, * sys.argv)
     #get the id of the selected div
-    def get_id():
-        text_id = printButton.cget('text')
-        need_id = text_id[text_id.find(':')+1:]
-        text_file = open("saves/selected_id.txt", "w")
-        text_file.writelines(need_id)
-        
+    def get_id():  
+        #make the input into a readable string
+        text_id = id_textBox.get("1.0","end-1c")
+        #text_id = printButton.cget('text')
+        #need_id = text_id[text_id.find(':')+1:]
+        #check if the id exists to edit is
+        id_exists =  True
+        for div in divs:
+            if text_id == div['id']:
+                id_exists = True
+                break
+            else:
+                id_exists = False
+            print(id_exists)
+        if text_id != "" and id_exists == True:
+            text_file = open("saves/selected_id.txt", "w")
+            text_file.writelines(text_id)
+        else:
+            error_label = tk.Label(root, text = "Enter a valid id!").pack()
+        id_textBox.delete("1.0","end-1c")
+
     text_file = open("saves/selected_file.txt", "r")
     get_file = text_file.readline()   
     from bs4 import BeautifulSoup as bs
@@ -25,7 +40,7 @@ def edit():
         with open("html_files/index.html") as fp:
             soup = bs(fp) 
     else:
-        with open("html_files/html"+get_file) as fp:
+        with open("html_files/html/"+get_file) as fp:
             soup = bs(fp) 
     #new window
     root = tk.Tk()
@@ -34,18 +49,24 @@ def edit():
     index_content = soup.title.text
     divs = soup.find_all('div')
     i = 0
+    #loop throught all the divs
     while i< len(divs):
         text_label = str
+        #if it finds a paragraph then add paragraph
         if "<p>" in str(divs[i].findChildren()):
             text_label = "paragraph: "+divs[i]['id']
         else:
             text_label = "error"
-        printButton = tk.Button(root, text="edit " + text_label, command=get_id)
-        printButton.pack()
+        #show labels for what can be edited
+        tk.Label(root, text=text_label).pack()
         i+=1
+    tk.Label(root, text="Choose tag to edit").pack()
+    id_textBox = tk.Text(root, height = 1, width = 16)
+    id_textBox.pack()
+    from edit_paragraph import edit_par
+    tk.Button(root, text="edit tags", command=lambda:[get_id(), edit_par()]).pack()
     from move_tags import move
-    printButton = tk.Button(root, text="move tags", command=move)
-    printButton.pack()
+    tk.Button(root, text="move tags", command=move).pack()
     tk.Button(root, text="Restart", command=restart_program).pack()
     tk.Button(root, text="preview", command=preview).pack()
     root.mainloop()
